test for existence before removing files

+++ src/split_lab_wav.py
@@ -122,9 +122,12 @@ def split_in(folder, split_phones):
         print "did", lab_fname
         # some cleaning now:
         os.remove(lab_fname)
-        os.remove(mfc_fname)
-        os.remove(wav_fname)
-        os.remove(raw_fname)
+        if os.path.isfile(mfc_name):
+            os.remove(mfc_fname)
+        if os.path.isfile(wav_name):
+            os.remove(wav_fname)
+        if os.path.isfile(raw_name):
+            os.remove(raw_fname)
         os.remove(phones_fname)
